requirements.txt
django
django-markdownx==4.0.7